Fix handling of concurrent requests in prompt-proto-service.
This is a cherry-pick of @dspeck1's experiments. Previously, work was
being assigned to pods on a best-effort basis, causing jobs to pile up
during scale-up instead of being assigned to the newly created pods.
The new spec forbids assigning work to a pod that is already busy
(though there may be still be a pile-up if the pod times out but
continues processing).